+++ addons/zylann.hterrain/tools/detail_editor/detail_editor.gd
@@ -4,6 +4,8 @@ extends Control
 const HTerrainData = preload("res://addons/zylann.hterrain/hterrain_data.gd")
 
 signal detail_selected(index)
+# Emitted when the tool added or removed a detail map
+signal detail_list_changed
 
 onready var _item_list = get_node("ItemList")
 onready var _edit_dialog = get_node("EditDetailDialog")
@@ -40,8 +42,14 @@ func _edit_detail(index):
 	_dialog_target = index
 	
 	var texture = _terrain.get_detail_texture(_dialog_target)
-	
-	_edit_dialog.set_params(texture)
+	var params = {
+		"texture": texture,
+		"bottom_shade": _terrain.get_detail_shader_param(_dialog_target, "u_bottom_ao"),
+		"global_map_tint_bottom": _terrain.get_detail_shader_param(_dialog_target, "u_globalmap_tint_bottom"),
+		"global_map_tint_top": _terrain.get_detail_shader_param(_dialog_target, "u_globalmap_tint_top")
+	}
+
+	_edit_dialog.set_params(params)
 	_edit_dialog.popup_centered_minsize()
 
 
@@ -66,8 +74,12 @@ func _on_EditDetailDialog_confirmed(params):
 	if _dialog_target == -1:
 		var data = _terrain.get_data()
 		index = data._edit_add_detail_map()
+		emit_signal("detail_list_changed")
 
 	_terrain.set_detail_texture(index, params.texture)
+	_terrain.set_detail_shader_param(index, "u_globalmap_tint_bottom", params.global_map_tint_bottom)
+	_terrain.set_detail_shader_param(index, "u_globalmap_tint_top", params.global_map_tint_top)
+	_terrain.set_detail_shader_param(index, "u_bottom_ao", params.bottom_shade)
 
 	if _dialog_target == -1:
 		_update_list()
@@ -81,6 +93,7 @@ func _on_ConfirmationDialog_confirmed():
 	var data = _terrain.get_data()
 	data._edit_remove_detail_map(_dialog_target)
 	_update_list()
+	emit_signal("detail_list_changed")
 
 
 func _on_ItemList_item_selected(index):